@@ -1,8 +1,7 @@
 import handler from '../../../lib/handler'
-import { getDetailedPost } from '../../../lib/getDetailed'
+import { getPost } from '../../../lib/getPosts'
 
-export default handler
-.get( async (req, res) => {
-  const detailedPost = await getDetailedPost(req.query.id);
-  res.json(detailedPost)
+export default handler.get(async (req, res) => {
+  const post = await getPost(req.query.id)
+  res.json(post)
 })
