@@ -1,6 +1,6 @@
 import handler from '../../../../lib/handler'
-import { getDetailedPost, getUserPosts } from '../../../../lib/getDetailed'
 import User from '../../../../models/user'
+import { addPost } from '../../../../lib/addPost';
 
 export default handler
 .get( async (req, res) => {
@@ -9,21 +9,13 @@ export default handler
 })
 .post( async (req, res) => {
   const { subject, content } = req.body;
-  const postUser = await User.findById(req.user);
+  const userId = req.user
+
+  const newPost = await addPost(userId, subject, content)
   
-  const newPost = {
-    user: postUser.id, 
-    subject, 
-    content, 
-    date_posted: Date.now()
-  }
-  
-  postUser.posts.push(newPost);
-  const postedPost = postUser.getChanges().$push.posts.$each[0]
-  
-  postUser.save()
-    .then( async user => res.json({ 
-      msg: 'New Post Added', 
-      post: await getDetailedPost(postedPost) 
-    }))
+  newPost.save((err, post) => {
+    if (err) return res.status(400).json(err)
+    
+    res.status(200).json(post)
+  })
 })
